requirements.txt
aniso8601==7.0.0
astroid==2.4.2
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.9
Flask-GraphQL==2.0.1
Flask-GraphQL-Auth==1.3.2
graphene==2.1.8
graphql-core==2.3.2
graphql-relay==2.0.1
graphql-server-core==1.2.0
isort==5.6.4
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pep8==1.7.1
promise==2.3
PyJWT==1.7.1
pylint==2.6.0
pylint-mccabe==0.1.3
pymodm==0.4.3
pymongo==3.11.1
python-dotenv==0.15.0
Rx==1.6.1
six==1.15.0
toml==0.10.2
Werkzeug==1.0.1
wrapt==1.12.1
